feat: 삭제되지 않은 마일스톤만 조회하도록 변경 및 기타 리뷰 수정사항 반영
- MilestoneRepository 에 findActiveMilestoneById 추가 CRUD 모두 삭제되지 않은 마일스톤에만 반영하도록 변경
- 코파일럿 리뷰사항 수정

+++ apps/spring-api/src/main/java/com/codesquad/issueTracker/milestone/dto/MilestoneListResponse.java
@@ -11,9 +11,9 @@ public record MilestoneListResponse(
         List<MilestoneResponse> milestones
 ) {
     public MilestoneListResponse(List<Milestone> milestoneList, int openMilestoneCount, int closedMilestoneCount){
-        this((Integer) (openMilestoneCount + closedMilestoneCount),
-                (Integer)openMilestoneCount,
-                (Integer)closedMilestoneCount,
+        this( (openMilestoneCount + closedMilestoneCount),
+                openMilestoneCount,
+                closedMilestoneCount,
                 milestoneList.stream().map(MilestoneResponse::new).toList());
     }
 }
